pyproject.toml
[project]
name = "job-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.4.0",
    "groq>=1.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
